dashboard/requirements.txt
dash>=2.14.0
dash-bootstrap-components>=1.5.0
plotly>=5.18.0
pandas>=2.1.0
pyarrow>=14.0.0
bcrypt>=4.1.0
flask-session>=0.5.0
openpyxl>=3.1.0
scikit-learn>=1.3.0
